requirements.txt
# Core dependencies
fastapi>=0.95.0
uvicorn>=0.22.0
python-dotenv>=1.0.0
pydantic>=2.0.0
jinja2>=3.1.2

# AI and LLM integrations
openai>=1.3.0
requests>=2.31.0
langchain>=0.0.335
langchain-community>=0.0.10

# Ollama integration for local models
ollama>=0.1.0

# Vector database
chromadb>=0.4.18
qdrant-client>=1.6.4
whisper-cpp>=1.0.0  # Optional, install separately
piper-tts>=0.0.1    # Optional, install separately
pillow>=9.5.0

# API clients
google-api-python-client>=2.86.0
wolframalpha>=5.0.0  # Optional

# Utils
numpy>=1.24.3
pandas>=2.0.1
tqdm>=4.65.0
